fix bank bug, make fish produce after they eat
explanation of bank bug: banks were checking if you had enough coins for your purchase for each one before deducting any of the coins, so if you had enough coins for x number of thing a OR y number of thing b but not both, it would complete transactions for both, resulting in you having a negative nuber of coins. now, it checks for thing a, then completes the transaction, then after you've been charged x coins it checks for thing b.