added 2 UI sounds, fixed background music bug in firefox
in firefox, it loads the previous setting for select options, then calls the onchange, which played the background music. additonally, unpausing played the background music. now it checks to see if music is playing in both functions.

note: UI sounds don't play in chrome (policy to protect user from spam)